Refactor unit tests to use Session across tests.
